Move ArbeidsgiverNotifikasjon to package producer.arbeidsgivernotifikasjon.domain
