+++ src/main/kotlin/no/nav/syfo/service/ArbeidsgiverNotifikasjonService.kt
@@ -3,7 +3,7 @@ package no.nav.syfo.service
 import no.nav.syfo.*
 import no.nav.syfo.producer.arbeidsgivernotifikasjon.ArbeidsgiverNotifikasjonProdusent
 import no.nav.syfo.db.domain.VarselType
-import no.nav.syfo.kafka.producers.dinesykmeldte.domain.ArbeidsgiverNotifikasjon
+import no.nav.syfo.producer.arbeidsgivernotifikasjon.domain.ArbeidsgiverNotifikasjon
 import java.time.LocalDateTime
 import java.util.*
